% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltab.R
\name{modeltab}
\alias{modeltab}
\title{Format regression estimates in a data frame}
\usage{
modeltab(
  ...,
  main = "b",
  aux = "se",
  add_stats = c("n", "df", "rsq"),
  conf_level = 0.95
)
}
\arguments{
\item{...}{Regression models (the results of \code{lm()}) to be made into a data frame.}

\item{main}{A character string indicating the main estimate to be shown in the data frame. By default, the point estimate. See \code{aux} for alternative estimates (\code{"ci"} cannot be used as the main estimate).}

\item{aux}{A character string indicating the secondary estimate to be shown in the row beneath the main estimate. By default, the standard error of the coefficient (\code{"se"}). Use \code{"t"} for the t-statistics, \code{"p"} for the p-value, or \code{"ci"} for the confidence interval instead of \code{"se"} if desired.}

\item{add_stats}{A character vector indicating model-level statistics to add to the end (last rows) of the table. Possible values are \code{"n"} (number of observations), \code{"df"} (degrees of freedom), \code{"rsq"} (R-squared), \code{"arsq"} (adjusted R-squared), \code{"rss"} (residual sum of squares), \code{"mse"} (mean squared error), and \code{"f"} (F-statistic). Name the values (e.g., \code{c("Degrees of freedom" = "df")}) to change the row's label (leftmost cell).}

\item{conf_level}{When \code{aux = "ci"}, a number in the range (0, 1) indicating the confidence level to be used in calculating the confidence interval. By default 0.95 (for a 95\% confidence interval).}
}
\value{
A data frame with the specified regression summary estimates.
}
\description{
\lifecycle{deprecated}

This function is deprecated and will be rewritten in a future version.
Support for this function is temporarily suspended while a new implementation
is being developed.
}
\details{
\code{modeltab()} was deprecated in version 0.0.7. This function is being
rewritten and will be reintroduced in a future version with an improved
implementation.
}
