% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabdown.R
\name{tabdown}
\alias{tabdown}
\title{Format and generate a Markdown-formatted table from a data frame}
\usage{
tabdown(.data, out, headings, blank_headings = TRUE, suppress_savemsg = FALSE)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., a tibble).}

\item{out}{A character string that specifies the path of the output .Md file. If missing (default), the result is merely printed to the console.}

\item{headings}{A character vector that specifies the column headers for the table. When not specified, the data frame's column names will be used instead. When a named character vector, column headers will be the values of the vector whose names correspond to a column name in the original data frame. When a column from the data frame is not represented in the named character vector, the column header will be blank, unless \code{blank_headings} is changed to \code{FALSE}}

\item{blank_headings}{A logical value dictating whether column headers can be empty. If \code{TRUE}, columns left unspecified by the character vector passed as \code{headings} are blank. If \code{FALSE}, unspecified column headers are filled using the original data frame's column names.}

\item{suppress_savemsg}{A logical value for whether to suppress the message that confirms (a) that the table was saved and (b) the absolute filepath of that save.}
}
\description{
\lifecycle{deprecated}

This function is deprecated and will be rewritten in a future version.
Support for this function is temporarily suspended while a new implementation
is being developed.
}
\details{
\code{tabdown()} was deprecated in version 0.0.7. This function is being
rewritten and will be reintroduced in a future version with an improved
implementation.
}
