% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtex.R
\name{tabtex}
\alias{tabtex}
\title{Format and generate a LaTeX-formatted table from a data frame}
\usage{
tabtex(
  .data,
  out = NULL,
  digits = 3L,
  headings,
  blank_headings = FALSE,
  special_left = TRUE
)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., a tibble).}

\item{out}{A character string that specifies the path of the output tex file.}

\item{digits}{A non-negative integer value of the desired number of digits to show after the decimal in the table.}

\item{headings}{A character vector that specifies the column headers for the table. When not specified, the data frame's column names will be used instead. When a named character vector, column headers will be the values of the vector whose names correspond to a column name in the original data frame.}

\item{blank_headings}{A logical value dictating whether column headers can be empty.}

\item{special_left}{A logical value indicating whether the leftmost column should be formatted separately from the other columns (\code{TRUE} by default).}
}
\value{
A character string containing LaTeX-formatted table code, or writes to file if \code{out} is specified.
}
\description{
Format and generate a LaTeX-formatted table from a data frame
}
