% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtex.R
\name{tabtex}
\alias{tabtex}
\title{Format and generate a LaTeX-formatted table from a data frame}
\usage{
tabtex(
  .data,
  out,
  title,
  label,
  width = "0.8\\\\linewidth",
  position = "!htbp",
  note,
  note_width,
  note_label = TRUE,
  long_negatives = TRUE,
  numbered = TRUE,
  special_left = TRUE
)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g., a tibble).}

\item{out}{A character string that specifies the path of the output tex file.}

\item{title}{A character string that specifies the title for the table.}

\item{label}{A character string that specifies the LaTeX label (e.g., "tab:table1") for the table to be used in LaTeX references.}

\item{width}{A character string specifying the width of the table (by default, 0.8\linewidth, 80\% of the line width).}

\item{position}{A character string containing only elements of "h", "t", "b", "p", and "!" that dictate where the table is placed in the LaTeX document.}

\item{note}{A character string containing the note to place underneath the table.}

\item{note_width}{A character string that specifies the width of the note. By default, the note will be the same width as the table itself (specified with \code{width}).}

\item{note_label}{A character string containing the label preceding the note. By default, "Note:" in italics (\textit{Note:}).}

\item{long_negatives}{A logical value indicating whether negative signs preceding numbers in cells should be converted to mathematical negative signs ("$-$") in the LaTeX output.}

\item{numbered}{A logical value indicating whether the table should be numbered in LaTeX (if \code{TRUE}, uses the \code{table} environment; if \code{FALSE}, uses the \verb{table*} environment instead).}

\item{special_left}{A logical value indicating whether the leftmost column should be formatted separately from the other columns (\code{TRUE} by default).}
}
\description{
Format and generate a LaTeX-formatted table from a data frame
}
\examples{
# create a LaTeX-formatted table to display the mtcars dataset
tabtex(mtcars)

# reformat the mtcars dataset and save it as a LaTeX table
mtcars
cars <- mtcars[, c("mpg", "hp", "wt", "am")]
cars$am[cars$am == 1] <- "Manual"
cars$am[cars$am == 0] <- "Automatic"
names(cars) <- c("MPG", "Horsepower", "Weight (1000 lbs)", "Transmission")

tabtex(cars, out = "simple_cars.tex")

# do similar reformatting with dplyr but add a note specifying that weight is in 1000s of pounds
library(dplyr)

cars <- mtcars \%>\%
  mutate(am = ifelse(am == 1, "Manual", "Automatic")) \%>\%
  select("MPG" = mpg, "Horsepower" = hp, "Weight" = wt, "Transmission" = am) \%>\%
  tabtex(out = "dplyr_cars.tex", note = "Weight is measured in thousands of pounds.")
}
